Add files via upload
Abstracted the idea of the initial priority for every thread, and refactored the idea of the initial thread's priority. I noticed that a thread which is putted in a mid-level position in scheduler's list has a better testing outcome. Therefore TCB*-priority is now initalised to 3(mid level ) instead of 0 (highest possible)